+++ go/src/main/java/tech/notpaper/go/controller/ResponseController.java
@@ -1,0 +1,10 @@
+package tech.notpaper.go.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/go/api")
+public class ResponseController {
+
+}
